requirements.txt
browser-use>=0.1.40
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
playwright>=1.44.0
python-dotenv>=1.0.0
schedule>=1.2.2
rich>=13.7.0
jinja2>=3.1.4
